+++ code/preprocessing/punctuation_remover.py
@@ -11,7 +11,7 @@ Created on Wed Sep 29 09:45:56 2021
 import string
 import re
 from code.preprocessing.preprocessor import Preprocessor
-from code.util import COLUMN_TWEET, COLUMN_PUNCTUATION, deEmojify
+from code.util import COLUMN_TWEET, COLUMN_PUNCTUATION 
 
 # removes punctuation from the original tweet
 # inspired by https://stackoverflow.com/a/45600350
@@ -29,6 +29,23 @@ class PunctuationRemover(Preprocessor):
         """
         # input column "tweet", new output column
         super().__init__([COLUMN_TWEET], COLUMN_PUNCTUATION)
+    
+    def deEmojify(text):
+        """
+        Custom function. Removes all Unicodes from a string
+        
+        Arguments
+        ---------
+        text: str
+            Input text
+        Returns
+        --------
+        str_de: str
+            Input text stripped off elements that cannot be encoded as ascii symbols    
+        """
+        str_en = text.encode("ascii", "ignore")
+        str_de = str_en.decode()
+        return str_de
     
     # set internal variables based on input columns
     def _set_variables(self, inputs):
@@ -58,7 +75,7 @@ class PunctuationRemover(Preprocessor):
         # replace punctuation with empty string
         column = inputs[0].str.replace(self._punctuation, "")
         # remove all emojis
-        column_demojified = [deEmojify(str(sent)) for sent in column]
+        column_demojified = [PunctuationRemover.deEmojify(str(sent)) for sent in column]
         # replace non-alphanumeric characters
         column_stripped = [re.sub('[^A-Za-z0-9\s]+', '', str(sent)) for sent in column_demojified]
         return column_stripped
